+++ lib/interpreter.h
@@ -36,17 +36,17 @@ constexpr short total_length = revf_offset + revf_stack_length;
 constexpr std::array<int, MEMORY_SIZE> PERSISTENT_MEMORY = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
 
 struct Ctx;
-using op_func = void (*)(Ctx *, int);
-void read(Ctx *ctx, int arg);
-void add(Ctx *ctx, int arg);
-void set(Ctx *ctx, int arg);
-void push(Ctx *ctx, int arg);
-void print(Ctx *ctx, int arg);
-void swap(Ctx *ctx, int arg);
-void grow(Ctx *ctx, int arg);
-void inp(Ctx *ctx, int arg);
-void jmpm(Ctx *ctx, int arg);
-void revf(Ctx *ctx, int arg);
+using op_func = void (*)(Ctx *, short);
+void read(Ctx *ctx, short arg);
+void add(Ctx *ctx, short arg);
+void set(Ctx *ctx, short arg);
+void push(Ctx *ctx, short arg);
+void print(Ctx *ctx, short arg);
+void swap(Ctx *ctx, short arg);
+void grow(Ctx *ctx, short arg);
+void inp(Ctx *ctx, short arg);
+void jmpm(Ctx *ctx, short arg);
+void revf(Ctx *ctx, short arg);
 
 struct Ctx
 {
@@ -75,7 +75,7 @@ struct Ctx
     };
 };
 
-inline void print_debug(Ctx *ctx, int chr_int, int last)
+inline void print_debug(Ctx *ctx, short chr_int, short last)
 {
     std::cout << separator;
     std::cout << "STACK\n| ";
@@ -135,7 +135,7 @@ inline std::array<int, MEMORY_SIZE> run(std::stringstream &_data, const bool deb
     ctx.pointer = data.data();
     short last = -1;
     TermiosRaw raw;
-    int chr_int;
+    short chr_int;
     const unsigned char *const end = data.data() + data.size();
     while (ctx.pointer < end || ctx.transposus_ok)
     {
